refactor/ move date_format_iso_to_ymd/1 to Utils.DatetimeProcessor
