+++ app/lib/kokuraex/services/datetime_function.ex
@@ -25,24 +25,6 @@ defmodule Kokuraex.Services.DatetimeFunction do
   def return_datetime({:error, _}), do: "0000-00-00 00:00(JST)"
 
   @doc """
-  ISO8601形式 "2021-01-01T00:00:00Z" を "2021-01-01" 日付文字列に置換する
-  """
-  @spec date_format_iso_to_ymd(String.t()) :: String.t()
-  def date_format_iso_to_ymd(iso_date_string) do
-    iso_date_string
-    |> DateTime.from_iso8601()
-    |> case do
-      {:ok, datetime, _} ->
-        datetime
-        |> DateTime.to_date()
-        |> Date.to_string()
-
-      {:error, _} ->
-        "0000-00-00"
-    end
-  end
-
-  @doc """
   現在のJST時刻を取得してISO8601形式で返す
   """
   @spec current_jst_datetime() :: String.t()
