+++ app/lib/kokuraex/services/news_function.ex
@@ -3,7 +3,7 @@ defmodule Kokuraex.Services.NewsFunction do
   Provides functions for handling news about Elixir topics.
   """
 
-  alias Kokuraex.Services.DatetimeFunction
+  alias Kokuraex.Utils.DatetimeProcessor
   alias Kokuraex.Utils.Cache
 
   defp _github_api_header() do
@@ -57,7 +57,7 @@ defmodule Kokuraex.Services.NewsFunction do
           created_at:
             information_json
             |> Map.get("created_at")
-            |> DatetimeFunction.date_format_iso_to_ymd(),
+            |> DatetimeProcessor.date_format_iso_to_ymd(),
           url: information_json |> Map.get("html_url"),
           body:
             information_json
